requirements.txt
# Multi-Platform Deep Research Agent Dependencies

# Core LangChain Framework
langchain>=0.1.0
langgraph>=0.0.20
langchain-core>=0.1.0

# LangChain Integrations
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Search API Libraries
exa-py>=1.0.0
tavily-python>=0.3.0
requests>=2.31.0

# MCP Server (SSE transport for remote deployment)
mcp>=1.0.0
starlette>=0.36.0
sse-starlette>=1.8.0
uvicorn>=0.25.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
